web: Move player stuff to Player module
