+++ web/packages/core/src/public/player/index.ts
@@ -1,0 +1,4 @@
+export * from "./flash";
+export * from "./player-element";
+export * from "./movie-metadata";
+export * from "./legacy";
