+++ web/packages/selfhosted/test/js_api/load.ts
@@ -1,7 +1,7 @@
 import { loadJsAPI, playAndMonitor } from "../utils.js";
 import { use } from "chai";
 import chaiHtml from "chai-html";
-import { PlayerElement } from "ruffle-core";
+import { Player } from "ruffle-core";
 
 use(chaiHtml);
 
@@ -12,7 +12,7 @@ describe("RufflePlayer.load", () => {
         const player = await browser.$("<ruffle-player>");
         await browser.execute(async (playerElement) => {
             // https://github.com/webdriverio/webdriverio/issues/6486
-            const player = playerElement as unknown as PlayerElement;
+            const player = playerElement as unknown as Player.PlayerElement;
             await player.load("/test_assets/example.swf");
         }, player);
         await playAndMonitor(browser, player);
