+++ web/packages/selfhosted/test/js_api/metadata.ts
@@ -1,7 +1,7 @@
 import { loadJsAPI } from "../utils.js";
 import { expect, use } from "chai";
 import chaiHtml from "chai-html";
-import { PlayerElement } from "ruffle-core";
+import { Player } from "ruffle-core";
 
 use(chaiHtml);
 
@@ -12,7 +12,7 @@ describe("RufflePlayer.metadata", () => {
         const player = await browser.$("<ruffle-player>");
         const metadata = await browser.execute(
             // https://github.com/webdriverio/webdriverio/issues/6486
-            (player) => (player as unknown as PlayerElement).metadata,
+            (player) => (player as unknown as Player.PlayerElement).metadata,
             player,
         );
         // [NA] Work around a chrome 87 bug where it's (somehow) adding extra data to this object
